Moved Helper to util package
