+++ src/edu/neu/csye6200/util/Helper.java
@@ -1,4 +1,4 @@
-package edu.neu.csye6200;
+package edu.neu.csye6200.util;
 
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
